Consolidate CEL test files, convert stdlib to Ginkgo, unify nesting
Section 3: Consolidate 13 CEL test files into 5 by merging small
overlapping files into their logical targets. Delete 8 source files.

Section 4: Convert all stdlib func Test... functions to Ginkgo
Describe/It/DescribeTable in regression_test.go, architecture_
application_test.go, architecture_removal_test.go, and
architecture_inplace_update_test.go.

Section 5: Unify multiple var _ = Describe blocks in each consolidated
file into a single Describe with Context sub-groupings for consistent
Describe -> Context -> It nesting.
